+++ angular-workspace/projects/ni/nimble-angular/table-column/enum-text/nimble-table-column-enum-text.directive.ts
@@ -2,9 +2,10 @@ import { Directive, ElementRef, Input, Renderer2 } from '@angular/core';
 import { type TableColumnEnumText, tableColumnEnumTextTag } from '@ni/nimble-components/dist/esm/table-column/enum-text';
 import { BooleanValueOrAttribute, NumberValueOrAttribute, toBooleanProperty, toNullableNumberProperty } from '@ni/nimble-angular/internal-utilities';
 import { NimbleTableColumnBaseDirective } from '@ni/nimble-angular/table-column';
-import type { MappingKeyType } from '@ni/nimble-components/dist/esm/table-column/enum-base/types';
+import { MappingKeyType } from '@ni/nimble-components/dist/esm/table-column/enum-base/types';
 
-export type { TableColumnEnumText, MappingKeyType };
+export { MappingKeyType };
+export type { TableColumnEnumText };
 export { tableColumnEnumTextTag };
 
 /**