+++ packages/nimble-components/src/table-column/number-text/tests/table-column-number-text.spec.ts
@@ -88,7 +88,7 @@ describe('TableColumnNumberText', () => {
             await connect();
             await waitForUpdatesAsync();
 
-            expect(pageObject.getRenderedCellContent(0, 0)).toBe('');
+            expect(pageObject.getRenderedCellTextContent(0, 0)).toBe('');
         });
     }
 
@@ -107,14 +107,14 @@ describe('TableColumnNumberText', () => {
         await connect();
         await waitForUpdatesAsync();
 
-        expect(pageObject.getRenderedCellContent(0, 0)).toBe('2.9');
-        expect(pageObject.getRenderedGroupHeaderContent(0)).toBe('2.9');
+        expect(pageObject.getRenderedCellTextContent(0, 0)).toBe('2.9');
+        expect(pageObject.getRenderedGroupHeaderTextContent(0)).toBe('2.9');
 
         columnInstances.column1.format = NumberTextFormat.roundToInteger;
         await waitForUpdatesAsync();
 
-        expect(pageObject.getRenderedCellContent(0, 0)).toBe('3');
-        expect(pageObject.getRenderedGroupHeaderContent(0)).toBe('3');
+        expect(pageObject.getRenderedCellTextContent(0, 0)).toBe('3');
+        expect(pageObject.getRenderedGroupHeaderTextContent(0)).toBe('3');
     });
 
     it('shows initial values', async () => {
@@ -122,9 +122,9 @@ describe('TableColumnNumberText', () => {
         await connect();
         await waitForUpdatesAsync();
 
-        expect(pageObject.getRenderedCellContent(0, 0)).toBe('2');
-        expect(pageObject.getRenderedGroupHeaderContent(0)).toBe('2');
-        expect(pageObject.getRenderedCellContent(0, 1)).toBe('-99');
+        expect(pageObject.getRenderedCellTextContent(0, 0)).toBe('2');
+        expect(pageObject.getRenderedGroupHeaderTextContent(0)).toBe('2');
+        expect(pageObject.getRenderedCellTextContent(0, 1)).toBe('-99');
     });
 
     it('changing fieldName updates display', async () => {
@@ -135,36 +135,36 @@ describe('TableColumnNumberText', () => {
         columnInstances.column1.fieldName = 'number2';
         await waitForUpdatesAsync();
 
-        expect(pageObject.getRenderedCellContent(0, 0)).toBe('-99');
-        expect(pageObject.getRenderedGroupHeaderContent(0)).toBe('-99');
+        expect(pageObject.getRenderedCellTextContent(0, 0)).toBe('-99');
+        expect(pageObject.getRenderedGroupHeaderTextContent(0)).toBe('-99');
     });
 
     it('changing data from value to null displays blank', async () => {
         await element.setData([{ number1: 10 }]);
         await connect();
         await waitForUpdatesAsync();
-        expect(pageObject.getRenderedCellContent(0, 0)).toBe('10');
-        expect(pageObject.getRenderedGroupHeaderContent(0)).toBe('10');
+        expect(pageObject.getRenderedCellTextContent(0, 0)).toBe('10');
+        expect(pageObject.getRenderedGroupHeaderTextContent(0)).toBe('10');
 
         await element.setData([{ number1: null }]);
         await waitForUpdatesAsync();
 
-        expect(pageObject.getRenderedCellContent(0, 0)).toBe('');
-        expect(pageObject.getRenderedGroupHeaderContent(0)).toBe('');
+        expect(pageObject.getRenderedCellTextContent(0, 0)).toBe('');
+        expect(pageObject.getRenderedGroupHeaderTextContent(0)).toBe('');
     });
 
     it('changing data from null to value displays value', async () => {
         await element.setData([{ number1: null }]);
         await connect();
         await waitForUpdatesAsync();
-        expect(pageObject.getRenderedCellContent(0, 0)).toBe('');
-        expect(pageObject.getRenderedGroupHeaderContent(0)).toBe('');
+        expect(pageObject.getRenderedCellTextContent(0, 0)).toBe('');
+        expect(pageObject.getRenderedGroupHeaderTextContent(0)).toBe('');
 
         await element.setData([{ number1: -16 }]);
         await waitForUpdatesAsync();
 
-        expect(pageObject.getRenderedCellContent(0, 0)).toBe('-16');
-        expect(pageObject.getRenderedGroupHeaderContent(0)).toBe('-16');
+        expect(pageObject.getRenderedCellTextContent(0, 0)).toBe('-16');
+        expect(pageObject.getRenderedGroupHeaderTextContent(0)).toBe('-16');
     });
 
     it('when no fieldName provided, nothing is displayed', async () => {
@@ -175,8 +175,8 @@ describe('TableColumnNumberText', () => {
         await element.setData([{ number1: 11 }]);
         await waitForUpdatesAsync();
 
-        expect(pageObject.getRenderedCellContent(0, 0)).toBe('');
-        expect(pageObject.getRenderedGroupHeaderContent(0)).toBe('');
+        expect(pageObject.getRenderedCellTextContent(0, 0)).toBe('');
+        expect(pageObject.getRenderedGroupHeaderTextContent(0)).toBe('');
     });
 
     describe('displays title when appropriate', () => {
@@ -277,6 +277,103 @@ describe('TableColumnNumberText', () => {
         });
     });
 
+    it('uses decimal-digits applied before connection', async () => {
+        columnInstances.column1.format = NumberTextFormat.decimal;
+        columnInstances.column1.decimalDigits = 4;
+        await element.setData([{ number1: 11.01234567 }]);
+        await connect();
+        await waitForUpdatesAsync();
+
+        expect(pageObject.getRenderedCellTextContent(0, 0)).toBe('11.0123');
+        expect(pageObject.getRenderedGroupHeaderTextContent(0)).toBe('11.0123');
+    });
+
+    describe('updating configuration after connection', () => {
+        beforeEach(async () => {
+            columnInstances.column1.format = NumberTextFormat.decimal;
+            await connect();
+            await element.setData([{ number1: 11 }]);
+            await waitForUpdatesAsync();
+        });
+
+        it('displays two decimal digits by default', () => {
+            expect(pageObject.getRenderedCellTextContent(0, 0)).toBe('11.00');
+            expect(pageObject.getRenderedGroupHeaderTextContent(0)).toBe(
+                '11.00'
+            );
+        });
+
+        it('updating decimal-digits updates rendered value', async () => {
+            columnInstances.column1.decimalDigits = 5;
+            await waitForUpdatesAsync();
+
+            expect(pageObject.getRenderedCellTextContent(0, 0)).toBe(
+                '11.00000'
+            );
+            expect(pageObject.getRenderedGroupHeaderTextContent(0)).toBe(
+                '11.00000'
+            );
+        });
+
+        it('updating decimal-digits to undefined uses two digits', async () => {
+            columnInstances.column1.decimalDigits = 5;
+            await waitForUpdatesAsync();
+            columnInstances.column1.decimalDigits = undefined;
+            await waitForUpdatesAsync();
+
+            expect(pageObject.getRenderedCellTextContent(0, 0)).toBe('11.00');
+            expect(pageObject.getRenderedGroupHeaderTextContent(0)).toBe(
+                '11.00'
+            );
+        });
+
+        it('setting an invalid decimal-digits value makes the column invalid', async () => {
+            columnInstances.column1.decimalDigits = -5;
+            await waitForUpdatesAsync();
+
+            expect(columnInstances.column1.checkValidity()).toBeFalse();
+            expect(
+                columnInstances.column1.validity.invalidDecimalDigits
+            ).toBeTrue();
+        });
+
+        it('changing format of invalid decimal column makes it valid', async () => {
+            columnInstances.column1.decimalDigits = -5;
+            await waitForUpdatesAsync();
+
+            expect(columnInstances.column1.checkValidity()).toBeFalse();
+            expect(
+                columnInstances.column1.validity.invalidDecimalDigits
+            ).toBeTrue();
+
+            columnInstances.column1.format = NumberTextFormat.default;
+            await waitForUpdatesAsync();
+
+            expect(columnInstances.column1.checkValidity()).toBeTrue();
+            expect(
+                columnInstances.column1.validity.invalidDecimalDigits
+            ).toBeFalse();
+        });
+
+        it('changing to a valid decimal-digits value makes an invalid column valid', async () => {
+            columnInstances.column1.decimalDigits = -5;
+            await waitForUpdatesAsync();
+
+            expect(columnInstances.column1.checkValidity()).toBeFalse();
+            expect(
+                columnInstances.column1.validity.invalidDecimalDigits
+            ).toBeTrue();
+
+            columnInstances.column1.decimalDigits = 1;
+            await waitForUpdatesAsync();
+
+            expect(columnInstances.column1.checkValidity()).toBeTrue();
+            expect(
+                columnInstances.column1.validity.invalidDecimalDigits
+            ).toBeFalse();
+        });
+    });
+
     const alignmentTestCases = [
         {
             name: 'with default format and default alignment',
@@ -311,6 +408,24 @@ describe('TableColumnNumberText', () => {
         {
             name: 'with roundToInteger format and right alignment',
             format: NumberTextFormat.roundToInteger,
+            configuredColumnAlignment: NumberTextAlignment.right,
+            expectedCellViewAlignment: TextCellViewBaseAlignment.right
+        },
+        {
+            name: 'with decimal format and default alignment',
+            format: NumberTextFormat.decimal,
+            configuredColumnAlignment: NumberTextAlignment.default,
+            expectedCellViewAlignment: TextCellViewBaseAlignment.right
+        },
+        {
+            name: 'with decimal format and left alignment',
+            format: NumberTextFormat.decimal,
+            configuredColumnAlignment: NumberTextAlignment.left,
+            expectedCellViewAlignment: TextCellViewBaseAlignment.left
+        },
+        {
+            name: 'with decimal format and right alignment',
+            format: NumberTextFormat.decimal,
             configuredColumnAlignment: NumberTextAlignment.right,
             expectedCellViewAlignment: TextCellViewBaseAlignment.right
         }